+++ jdaviz/configs/default/plugins/plot_options/plot_options.py
@@ -18,7 +18,6 @@ from glue.viewers.image.composite_array import COLOR_CONVERTER
 from glue_jupyter.bqplot.image.state import BqplotImageLayerState
 from glue_jupyter.common.toolbar_vuetify import read_icon
 
-from jdaviz.components.toolbar_nested import NestedJupyterToolbar
 from jdaviz.core.registries import tray_registry
 from jdaviz.core.template_mixin import (PluginTemplateMixin, ViewerSelect, LayerSelect,
                                         PlotOptionsSyncState, Plot,
@@ -465,9 +464,7 @@ class PlotOptions(PluginTemplateMixin):
         self.stretch_histogram = Plot(self, viewer_type='histogram')
         # Add the stretch bounds tool to the default Plot viewer.
         self.stretch_histogram.tools_nested.append(["jdaviz:stretch_bounds"])
-        self.stretch_histogram.toolbar = NestedJupyterToolbar(self.stretch_histogram.viewer,
-                                                              self.stretch_histogram.tools_nested,
-                                                              ["jdaviz:stretch_bounds"])
+        self.stretch_histogram._initialize_toolbar(["jdaviz:stretch_bounds"])
 
         # NOTE: this is a current workaround so the histogram viewer doesn't crash when replacing
         # data.  Note also that passing x=[0] fails on SOME machines, so we'll pass [0, 1] instead